REF: major code reorg, + tying up loose ends as we dig into 0.4.1

+++ udder/src/main/java/com/coillighting/udder/mix/Mixable.java
@@ -31,7 +31,7 @@ public interface Mixable extends Effect {
      *
      *  This is your layer's volume knob, a.k.a. its submaster.
      */
-    public float getLevel();
-    public void setLevel(float level);
+    public double getLevel();
+    public void setLevel(double level);
 
 }
